pulled JSONEditor extensions into separate files
